fix(#35): 채팅방 참여 모달 수정
